Retry failed etcd handlers
Before, if a handler on an etcd subscriber failed (returned an error) we
would ignore it. In the context of managing PGBouncer, this meant if
PGBouncer wasn't up, or went down missing a host change, then it would
never get updated.

This change introduces indefinite retry of etcd subscriber handlers to
repeatedly try getting PGBouncer to point at the correct host.