Integration test for supervise behaviour
Create a test that verifies supervise behaviour, for both the proxy and
cluster, against a real three node Postgres cluster. The test starts by
booting the cluster, injecting the debian package by mounting the
project directory.

Once booted, we verify that the PGBouncer on the cluster machines are
updated to point at the current master node. After this, we test a
migration and confirm that PGBouncers are updated to point at the
latest master node.