Create migration API
We need to expose three API endpoints to perform a zero-downtime
migration. These are:

- POST /pause?timeout&expiry
- POST /resume
- POST /migrate

These respectively provide the ability to pause PGBouncer, resume
PGBouncer and issue a migration to pacemaker, causing the current
primary to be migrated to the sync node.

This will form the API that the migration flow will use to switch
Postgres primaries without clients seeing any interruption of service.