Create migration API
Expose a grpc API that will enable a migration flow where we can switch
Postgres primaries without clients seeing any interruption of service.

The API endpoints will provide the ability to pause PGBouncer, resume
PGBouncer and issue a migration to pacemaker, causing the current
primary to be migrated to the sync node.