Use exec for PGBouncer actions
When passing contexts to the PGBouncer functions, we want to ensure we
don't terminate the context while we still have a connection open, as
PGBouncer will rollback the command. This results in actions like PAUSE
being undone even when the calling code believes them to have been
successful.

To avoid this, use the exec sql.DB family that throws away the returned
rows, better handling the closing of connections.