Cib -> Pacemaker and add context params
Previously we had the cib executor configure a set timeout, but this is
inflexible and doesn't benefit from context chaining where the parent
context is cancelled and resources are cleaned up in the dependents.

This change also modifies the terminology so that we don't call the
interface around pacemaker a cib, instead providing a general Pacemaker
API. This gets slightly hairy with naming as the word pacemaker is taken
by the module, but we get away with this by terming the Pacemaker
instance a crm (which is more appropriate than cib used to be).