Adjust migration flow to expect IP addresses
Now that we push IP addresses instead of node names to etcd, the
migration flow (which before expected node names) needs to be adjusted.
This change simply responds with the new master address when issuing a
migration, and we wait for the address to be sent to etcd before
confirming that the migration is complete.