Allow config to set pgbouncer credentials
We need to allow the cluster manager to provide credentials to pgbouncer
when connecting, as we don't have trust mode in production and we won't
be running with the same user as the pgbouncer socket given we'll need
to be root for migrations.

This change creates a simple AuthorizedExecutor struct that knows how to
run commands against the pgbouncer socket. Instead of the previous
approach where we load settings from the pgbouncer config file, we
simplify the executor by specifying the connection parameters up-front.

Making changes to the configuration meant it made more sense to
provision a config.toml into the test container, and specify our desired
config via that file as opposed to cli flags in the start-cluster
script.