Rename dockerfiles
Create a single image called pgsql-cluster-manager-base which has all
the dependencies required to run a full integration test environment,
along with building the package.

Layer CircleCI specific configuration on-top of the base image so that
we play nicely in CI.

Finally configure goreleaser to push alpine linux images with just the
binary for consuming from kubernetes.