Fix docker stonith resource
This change modifies the docker stonith resource to fix some bugs that
were introduced some time back. The resource was broken when trying to
turn off other nodes, which wasn't noticed in tests as we don't directly
expect against this, but meant the sandbox environment was less useful
for realistic testing.

We can test this change by executing the following commands, and
verifying that the pg03 container has been killed:

```bash
cd docker/postgres-member
./start
$ fence_pcmk -n pg03 -o off
```