Fix cmd package bugs
Ensure we present the config hash in the correct type for go-kit to log
the value. Fixup our config parsing to ensure our bytes are interpreted
as toml.

More significantly, we were using the uname attribute from pacemaker
cib, which we can't then pass to the resolve method to transform into an
IP address. This properly breaks things- this commit fixes them!