Ensure any assertions happen within ginkgo
Change the integration harnesses to not return errors, but rely on
gomega matchers to bail out of tests if things go wrong. We then need to
ensure we're within a ginkgo context if we're going to panic with a
matcher error, so start and cleanup the processes within a suite tear
up/down hook.