Fix-up postgres auth for unit tests
Ensure circleci runs our unit tests as a non-root user, as any pgbouncer
processes spawned by root will refuse to boot. Also ensure the postgres
user in circle has a password configured, and that the pgbouncer test
harness provides a mechanism to provision that password into the
userlist, as it will use this to auth against the target postgres.